Added IMU driver things from the driver Luca found on the Internet

+++ main_board/mb_working_jonas_sd_card/Core/Inc/util.h
@@ -41,8 +41,8 @@ typedef enum {
 
 /* IMU data */
 typedef struct {
-	int32_t gyro_x, gyro_y, gyro_z;
-	int32_t acc_x, acc_y, acc_z;
+	int16_t gyro_x, gyro_y, gyro_z;
+	int16_t acc_x, acc_y, acc_z;
 	timestamp_t ts;
 } imu_data_t;
